Remove use of `.device` from cirq-aqt (#4812)
Second part of #4744 

As far as I can tell there was just one use of `.device` in cirq-aqt in the form of an assertion that the attached device to a circuit was in fact a `cirq.IonDevice` where no call to validate ever took place. Also somewhat confusingly the number of qubits used in simulation would be dictated by the device qubit count instead of the circuit qubit count which could lead to some confusion on the users end.

I've gone ahead and just removed the use of device to revert this confusing behavior (which does make this a breaking change).