Use CIRCUIT_SERIALIZER as default serializer for quantum engine
CIRCUIT_SERIALIZER can serialize gates from any other gateset, uses a
more compact proto serialization that doesn't embed redundant string
literals, and can serialize more complicated constructs such as
CircuitOperation. It is meant to provide serialization separate from the
notion of which gates are supported by a particular device, two concepts
which were previously conflated. Having this as a default is a big UX
improvement when using quantum engine, as the gate_set argument can be
omitted on most calls. The gate_set argument could be removed in the
future, as specifying the serializer once on EngineContext, along with
the protobuf version, is sufficient to customize the serialization when
needed.