Allow repetitions to be parameterized (#5043)
Since recursive parameter resolution is now working (#5033) we can do this now.

The biggest caveat with this code is that params are floats, and repetitions must be an integer. I added a new type IntParam for the `repetitions` field itself, but it's still possible for the resolver to put a float value there. I added a runtime check for that. It may make sense to allow floats if they're really close to an integer, but I didn't do that here yet.

Closes #3266