Extract SV/DM/CliffordSimState from simulators (#5260)
First part of #5244, so that we don't have two variants of "XYZSimulatorState".

@95-martin-orion 

As mentioned in a comment to that issue, these sim states aren't exposed to the end user (except CliffordState), and end up being an unnecessary layer that wraps the ActOnArgs used in the simulation. This PR removes them from the simulation, and instead uses the ActOnArgs (specifically the OperationTarget[ActOnArgs]) directly. This allows us simply to deprecate and remove those classes in the future. (For Clifford, we also remove it from the simulation logic, and only create the CliffordState object in the public functions that expose it).

There are two distinct parts to this PR. The first was to move some of the shared logic between ActOnArgs and ActOnArgsContainer down a level in the hierarchy to OperationTarget. (*Primarily* this was to give a common `qubit_map` property rather than having to create one from the `qubits` each time it was required subsequently.

The second part is the migration from using the existing XYZSimState classes to using the ActOnArgs directly. The root of this is in `StepResultBase._simulator_state`. Previously each XYZStepResult defined its own `_simulator_state` property, that created the XYZSimState we're trying to extract. This PR deletes those and replaces them with the new `StepResultBase._simulator_state`, which returns the `OperationTarget[TActOnArgs]` instead. That bubbles up into `_final_simulator_state`, and the various functions in the XYZStepResults and XYZTrialResults have been modified to account for that change.

Finally, since the simulators are no longer bound to these `XYZSimulatorStates` anymore, `TSimulatorState` is removed from the type parameters everywhere possible.