Add native gates for IonQ
When circuits are submit, autodetect if only native gates are included, in which case the lang selection changes to 'native'.

This also changes the default from 'json' to 'qis' (a newer alias of the same language).

For now, it is not possible to compile to the native gates, but in a future change, we are hoping to add a CompilationTargetGatset for our
native gates.