Removing GateSpecification.valid_targets and deprecating some target types (#5376)
The original DeviceSpecification was designed to be somewhat future-proof, so it contains fields which do not pertain to hardware at the time. Some assumptions remain overly general for our hardware today and in the near future. Specifically, we can make the following assumptions about our devices:

* All valid qubit pairs work for all two-qubit gates.
* All valid qubits work for all single-qubit gates.
* Measurement gate can always be applied to all subset of qubits.

Based on these assumptions, `SUBSET_PERMUTATION` target type and the newly introduced `GateSpecification.valid_targets` are no longer necessary.

This change also removes the unused ASYMMETRIC target type.

As device requirements change, new fields can be added in the future and existing fields deprecated, and this can be done in a backward-compatible way as long as the server fills both deprecated and new fields until deprecated fields are removed. This provides the flexibility to keep our proto surface small, which makes serialization/deserialization logic simpler to maintain.

**Notes for reviewers**:
* `GateSpecification` was recently introduced and has never been sent over the wires, so it's OK to change its fields directly.
* The description of assumptions about 1q, 2q, and measurement targets are moved to the proto.
* The proto can be made simpler by replacing `TargetSet` with a less general `qubit_pairs`, which is sufficient to satisfy the requirements today. I would prefer this route as we have the flexibility to generalize later on if we'd like. WDYT?

@dstrain115 @maffoo