Fix #5773: guard against user setting git templateDir
If the user has a ~/.gitconfig file where init.templateDir is set,
then some of the tests in bash_scripts_test.py may fail because the
settings may cause the commands in the script to produce output that
doesn't match expectations. The changes here pass an explicit argument
to `git init` commands to override anything the user may have set in
their ~/.gitconfig file.