Improve solution based on feedback from @pavoljuhas
This solves the problem in a different way, using the solution
suggested by @pavoljuhas in comments on the PR. The new approach uses
environment variables GIT_CONFIG_SYSTEM and GIT_CONFIG_GLOBAL to
override per-user configuration in ~/.gitconfig. This is a better
solution because it makes git completely ignore the user's
configuration, rather than only addressing the single template dir
problem of my original solution.