Add {Frozen}Circuit.from_moments to construct circuit by moments.
In our internal code almost all circuits are constructed with explicit
moments rather than using an insertion strategy because we want to
control the alignment of ops. Here we add a `from_moments` classmethod
on `Circuit` and `FrozenCircuit` which takes any number of `OP_TREE`
args and converts each one to a `Moment` in the resulting circuit. This
gives a convenient way to construct circuits without cluttering the code
with multiple calls to `Moment`.