DROPME - temporarily upgrade to numpy-1.24.2
Temporarily upgrade to numpy-1.24.2 to enable CI testing of #5991.

Restrict CI-tests to Python 3.8 only which permits installation of
numpy-1.24.2 and numba.  Note numba is not yet compatible with
numpy-1.24 and its import raises SystemError exception
SystemError: initialization of _internal failed without raising an exception

Here we also add intentionally failing CI check to ensure this commit
is not accidentally merged to master.

Please revert this commit before merging the pull-request.