6097: Exploratory - Migrate from default Python serialization to orjson
This commit is part of an exploratory migration from the default Python
serialization system to the 'orjson' library. Changes include the addition of
the 'orjson' import in the json_serialization.py module and subsequent use of
the 'orjson' functions. In addition, 'orjson' has been added to the
cirq-core/requirements.txt file.