Make qubits a memoized property of Moment
Moment already has a _qubit_to_op dict. Maintaining a separate _qubits frozenset is inefficient when constructing circuits, as it needs copied each time an op is added to a moment.

In particular, for wide moments, this change seems to result in about a 3x speedup. On my laptop, creating a moment with X gates on 10_000 qubits takes 4s before this change, and 1.3s after.