Fix order preservation in align_right
Order was not preserved for non-commutable measurements and controls in the same moment. Reversing the circuit's moments did not reverse the ops within those moments. The subsequent `align_left` could split those ops into separate moments. Then reversing the moments again would have the end result of reversing the order of those operations in the final circuit.

This PR makes sure to reverse the order of ops within the moment prior to the call to `align_left`.