Add support for conditional blocks in README.rst
Due to differences in reStructuredText support between GitHub and
PyPI, some basic things like centering images are impossible to do
purely in RST in such a way that it works on both GitHub and
PyPI.:(This is not the case with Markdown; both PyPI and GitHub
support some limited HTML inside Markdown, and that makes it possible
to have the same Markdown README file work on both PyPI and GitHub.)

Since we are sticking with RST format for the top-level module README
files, I wanted to have a way to create an attractive project README
file for both GitHub and PyPI without having to keep two separate
files. However, that meant solving the following problem.

PyPI supports RST's `.. class::` directive, but GitHub does not. PyPI
also respects `:align:` attributes for images, but GitHub does not.
Consequently, the only way to center text or images in RST for GitHub
is to use raw HTML via the RST `.. raw::` directive, but it turns out
that PyPI not only _doesn't support `raw`_ – it actually _throws an
error and refuses your upload_ if your README file contains `.. raw::`
constructs.

What to do? Well, we _can_ use `.. raw:: html` as long as we remove
those constructs in the version of the file we save in the
distribution for PyPI. Further, since GitHub simply ignores or strips
out RST directives like `.. class::`, we can leave _those_ constructs
in the file that is sent to PYPI, and make use of them to do what we
can't do in GitHub purely in RST. So, in summary:

1. Make the unmodified README.rst file work for GitHub, because that's
   what people will see when they visit the repo on GitHub.

2. Filter the README.rst file when producing distributions for PyPI,
   to remove content delimited by specific start and end markers that
   indicate parts that are valid on GitHub but not PyPI.

3. Use normal RST constructs that work on PyPI but not GitHub to
   accomplish what we can't do for GitHub without using raw HTML.

That's the approach is used in the new updated top-level README.rst
file for Cirq, and it is enabled by this small modification to the
`setup.py` file.