Add manually-triggered debugging workflow
This workflow is designed to let you run pytest with various debugging
options. It's meant to be executed manually using "Run workflow"on
https://github.com/quantumlib/Cirq/actions/workflows/debug.yaml

Clicking the "Run workflow" button there will make GitHub present a
form interface with various options for the characteristics of the
run. Options include:

- give it a specific test to run (in the form of a path to a test file)
- set the number of repetitions of the test
- set the python version to use
- set the random seed
- turn on verbose logging

This workflow was developed as an aid to debugging issue #6906,
involving curve fitting that failed on occasion. This workflow is
general and I think it will be useful for other debugging needs.