Fix issue #6906 – flaky curve fitting tests
The test function ‘test_calibrate_z_phases()‘ in the file
‘cirq-core/cirq/experiments/z_phase_calibration_test.py‘ eventually
ends up calling `fit_exponential_decays()` in `xeb_fitting.py`. That
function uses SciPy's `curve_fit()` to fit an exponential function to
data. For some unlucky combination of random seed and problem cases,
`curve_fit()` can fail to find a solution, and raises `RuntimeError`
when that happens.

This change wraps the call to `curve_fit()` with a try-catch for
`RuntimeError`; when it happens, the wrapper calls `curve_fit()` one
more time with an explicit setting of higher max function evaluations.
If that fails too, then the resulting new `RuntimeError` will not be
caught.

This is admittedly not a true solution to the problem. In my testing,
this prevents the original CI error (with a particular random seed),
but it's probably the case that the second run of `curve_fit()` could
get unlucky again. Still, this may be good enough for most cases in
practice.

If we continue to see failures, we should explore a more sophisticated
solution to this.