Revert to simply running curve_fit with a higher maxfev
Discussions during code review resulted in the decision to not do a
try-fail-retry approach, and instead simply call `curve_fit()` once
but with an explicit, high value of `maxfev`.

This solution also work on the original case. To check, here is the
command that fails without this change and succeeds with the change:
(on a Debian-based Linux system, Python 3.11, NumPy 1.24, and SciPy
1.15.0):

```bash
./check/pytest -v --randomly-seed=3258636985 \
 cirq-core/cirq/experiments/z_phase_calibration_test.py::test_calibrate_z_phases_no_options
```

The above should be executed from the top of the Cirq source tree.