Add a new sweep metadata from/to proto approach (#6925)
Continue https://github.com/quantumlib/Cirq/pull/6869

This PR added a new path to encode and decode metadata to this
corresponding proto. Notice this is a side path if the metadata in sweep
is not encoded as `Metadata` class, the code path is still old and same
for `from_proto` case. If the proto does not contain `metadata` proto
field, the code still convert back as the old style