Use encoding UTF-8 when opening files
In many places, `open()` is called without providing an `encoding`
parameter. Turns out this can cause failures on Windows for files
containing, e.g., emoji characters in UTF-8 encoding, because the
default character encoding on Windows is not necessarily UTF-8.

Since all the files being read in the context of dev_tools/ scripts
can be assumed to be UTF-8, it seems sensible to use
`encoding="UTF-8"` in all calls to `open()`.