Update .markdownlintrc
After discovering the Google markdown style guide at
https://google.github.io/styleguide/docguide/style.html, it seemed
sensible to make our markdownlint configuration try to be as close to
the Google style as possible while still being compatible with
GitHub-flavored Markdown. This new config file tries to achieve that.

This update also changes the file format to JSONC. I was unable to
make the ini-style syntax work for the `allowed_elements` list under
`html`, and after spending far too long trying to figure out why, gave
up after realizing that such a waste of time might hit the next person
trying to edit the file. (Markdownlint ignores the `.markdownlintrc`
file if it contains a syntax error, but it _doesn't report what the
error is_, leaving it up to you to guess what's wrong. That's just
asinine. By contrast, when using JSONC format for `.markdownlintrc`,
it will report errors.)