Use encoding UTF-8 when opening files (#6989)
* Use encoding UTF-8 when opening files

In many places, `open()` is called without providing an `encoding`
parameter. Turns out this can cause failures on Windows for files
containing, e.g., emoji characters in UTF-8 encoding, because the
default character encoding on Windows is not necessarily UTF-8.

Since all the files being read in the context of dev_tools/ scripts
can be assumed to be UTF-8, it seems sensible to use
`encoding="UTF-8"` in all calls to `open()`.

* Remove dead code

[Per comment by
@pavol](https://github.com/quantumlib/Cirq/pull/6989#discussion_r1932797875)
on PR #6989, removing function `_get_version()` because it's dead code.

* Revert a couple of places where encoding is not needed

Per [comments by
@pavol](https://github.com/quantumlib/Cirq/pull/6989#pullrequestreview-2579284264)
on PR #6989, a couple of places didn't need the addition of an
`encoding` attribute on calls to `open()`.

* Remove unused import flagged by pylint

* Fix code style warnings from linter

* Fix incorrect line breakage

My editor said to break it, but Cirq's format settings say don't.
¯\_(ツ)_/¯