Update .markdownlintrc (#6990)
* Update .markdownlintrc

After discovering the Google markdown style guide at
https://google.github.io/styleguide/docguide/style.html, it seemed
sensible to make our markdownlint configuration try to be as close to
the Google style as possible while still being compatible with
GitHub-flavored Markdown. This new config file tries to achieve that.

This update also changes the file format to JSONC. I was unable to
make the ini-style syntax work for the `allowed_elements` list under
`html`, and after spending far too long trying to figure out why, gave
up after realizing that such a waste of time might hit the next person
trying to edit the file. (Markdownlint ignores the `.markdownlintrc`
file if it contains a syntax error, but it _doesn't report what the
error is_, leaving it up to you to guess what's wrong. That's just
asinine. By contrast, when using JSONC format for `.markdownlintrc`,
it will report errors.)

* Add first-line-heading

This adds `first-line-heading` setting, to bring another behavior
closer to Google Markdown style.

I also moved the two headings-related settings next to each other, to
organize the content a little bit better.

* Add $schema keyword

Turns out there's a JSON schema for this file format.

* Start top comment on first line of file

Emacs was not noticing the -*- jsonc -*- mode setting.

Despite the suspicious look of having a comment right after the
opening `{` brace, I tested it with the markdownlint program and it
does read the file without complaint.

* Add settings for Google list style guidelines

Embarrassingly, I only now read the section on list spacing in
https://google.github.io/styleguide/docguide/style.html#lists