Fix links to bib records & reposition recontributors number
The previous links for the bibliographic records went to Cirq 1.2 in
Zenodo. There doesn’t seem to be a way to get Zenodo to produce bibtex
for the latest version of a record (only specific versions of a
record), so I ended up using doi.org, but that one doesn't have a way
to produce MarcXML. So I removed that format, leaving the other two.

This also moves the GitHub contributors badge/count, per
recommendation by Pavol in the review of this PR.