Overhaul top-level README and convert to Markdown (#6988)
* Overhaul top-level README and convert to Markdown

This is a major overhaul of the main Cirq README file. Much of the
content is rewritten, and there are numerous additions and
enhancements all over. The changes include but are not limited to:

- Updates and corrections
- Addition of new sections
- Centering and resizing the logo for a more professional look
- Addition of GitHub badges for extra "pizzazz"
- Addition of pointers to related Quantumlib software
- Addition of a table of contents
- Improvements to info about how to cite Cirq
- Conversion to Markdown format to allow formatting that is
  compatible with both GitHub and PyPI

This is a follow-on to now-closed PRs #6903 and #6901, which concerned
making roughly the same changes while keeping with the
reStructuredText format of the previous README file. Switching to
Markdown allows use of some simple raw HTML to achieving formatting
that is not possible in pure Markdown, and still do so in a way that
works on both GitHub and PyPI.

* Adapt setup.py to work with new Markdown-format README

* Update to reference README.md instead of README.rst

* Use md reference links for better plain-text readability

No content edits (except possibly one or two trivial changes of
wording). This changes some links to use Markdown "reference" link
syntax, to make some of the source text more readable.

* Fix list indentation to follow Google Markdown style

* Fix links to bib records & reposition recontributors number

The previous links for the bibliographic records went to Cirq 1.2 in
Zenodo. There doesn’t seem to be a way to get Zenodo to produce bibtex
for the latest version of a record (only specific versions of a
record), so I ended up using doi.org, but that one doesn't have a way
to produce MarcXML. So I removed that format, leaving the other two.

This also moves the GitHub contributors badge/count, per
recommendation by Pavol in the review of this PR.

* Remove README.rst

* Add link target for #how-to-cite

This is so that existing links go to same section despite the renamed
heading.