Add workflow to save GitHub repo statistics
This saves GitHub repo statistics so that they can be analyzed for
longer durations than GitHub's 14-day limit. The workflow runs once a
day and saves the day's statistics in files in a branch of the git
repository. Data files accumulate over time. After every run, it
generates a report based on the cummulative data. The GitHub workflow
summary page has a link to this report.