Replace uses of word "secret" with "hidden"
The code scanning tools like CodeQL are not always able to distinguish
real problems from coincidences. One such situations is the use of the
word "secret" in some of the Cirq example programs. CodeQL is flagging
some of the code as printing plain-text secrets. The examples actually
_are_ about secrets, but they're examples, and what's being printed in
these cases is not actual passwords. Nevertheless, the simplest way to
avoid the problem (and possibly similar false-positives with other
security scanning software) is to use the word "hidden" instead of the
word "secret". It seems to work well enough in these contexts (IMHO).

This resolves the following code scanning alerts:

- https://github.com/quantumlib/Cirq/security/code-scanning/90
- https://github.com/quantumlib/Cirq/security/code-scanning/91
- https://github.com/quantumlib/Cirq/security/code-scanning/92