Use variables to hold important values
To avoid writing actual duration values in multiple places (and thus
face the risk of not updating some of them if we decide to change
values in the future), this now uses environment variables to store
the important values.

This also provides an easier way of writing the lists of exempt
labels. The stale handler action needs them as a comma-separated
string, and this is a bit of a PITA to edit when we make changes. We
can make it easier on ourselves in the future by using a list format
instead.