Fix "ambiguous redirect" error in check/shellcheck (#7078)
* Fix "ambiguous redirect" error in check/shellcheck

On my system, running `check/shellcheck` results in an error:

```
check/shellcheck: line 33: 0: ambiguous redirect
```

The reason concerned the use of the shell `<(...)` construct. It seems
that with some combination of user Bash options, versions of Bash, and
possibly other factors, the output is not compatible with using a `<`
redirection; however, I was unable to figure out exactly why. I
ultimately resolved it by using a slightly different syntax for the
git/pipeline output and variable assignment.

* Fix up formatting

1. Get rid of unnecessary backslash, per review comment.
2. Dedent the subsequent pipe commands to make things more clear.
3. Put the closing parens on a separate line, for style consistency.