Add a .editorconfig file (#7081)
* Resolve #6998: add a .editorconfig file

This adds settings based on some common-sense values and the Cirq
project's current conventions, such as line length and use of spaces
instead of tabs.

* Add explanation for .md setting for trailing spaces

* Take out exception to trailing spaces rule for .md files

Per discussion in the [review
comments](https://github.com/quantumlib/Cirq/pull/7081#discussion_r1968196913),
the feeling is that it's better to avoid trailing spaces because they
are bad for several reasons.