Update & improve workflow for stale issues & PRs (#7045)
* Fix labels & update messages

1. The previous settings were incorrect. The value of
   `stale-issue-label` should be `status/going-stale`; it is the value
   of `close-issue-label` that should be `status/stale`.

2. Added a couple of labels that would exempt PRs from staleness.

3. Per discussion on issue #6866, the message should be more clear
   that users can easily reopen the issue or PR by commenting on it.

* Fix text per review comments

This makes corrections to the "close" messages per
[review comments by @NoureldinYosri](https://github.com/quantumlib/Cirq/pull/7045#discussion_r1947415434).

* Use variables to hold important values

To avoid writing actual duration values in multiple places (and thus
face the risk of not updating some of them if we decide to change
values in the future), this now uses environment variables to store
the important values.

This also provides an easier way of writing the lists of exempt
labels. The stale handler action needs them as a comma-separated
string, and this is a bit of a PITA to edit when we make changes. We
can make it easier on ourselves in the future by using a list format
instead.

* Use a single label (status/stale) for all cases

Per [discussion in
review](https://github.com/quantumlib/Cirq/pull/7045#issuecomment-2649311401),
we decided we don't need separate labels `status/going-stale` and
`status/stale`. When an issue or PR is closed due to staleness, the
comment left by the workflow will make it clear why it was closed. We
can simplify things by using a single label.

* Change workflow name to be more like others

More of the workflow names in Cirq are, well, names, rather than
statements about what they do.

* Add workflow & job permissions, per best practices

* Skip the concurrency block to simplify this

Per comments on review, this workflow is not really in need of
guarding against concurrent execution. We can simplify it by removing
unnecessary bits like this.