Optimize the HHL algorithm using amplitude amplification.
I changed the example to simulate two circuits: the first one is the
previous HHL circuit in the example without amplitude amplification,
and the second one is the HHL circuit with amplitude amplification.

The example demonstrates that amplitude amplification obtains an estimate
faster (using fewer total calls to simulate.run).